requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
starlette==0.27.0

# LangGraph for workflow orchestration
langgraph==0.0.40
langchain-core==0.1.0

# Neo4j database
neo4j==5.15.0

# HTTP client for Ollama
aiohttp==3.9.1
aiosignal==1.3.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.0

# Vector search and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
huggingface-hub==0.19.0
transformers==4.35.0

# Data processing
numpy==1.26.0
scipy==1.11.0
scikit-learn==1.3.0
python-multipart==0.0.6
pandas==2.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.0
requests==2.31.0

# Optional: For production deployment
gunicorn==21.2.0

# Performance monitoring
psutil==5.9.0
